Move to checks dir
